Move the debug flag to its own package
